# backend/tests/catalog/openapi_snapshot.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Medical Prescription API",
    "version": "v1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/nudges/call-doctor": {
      "post": {
        "operationId": "initiate_doctor_call_api_v1_nudges_call_doctor_post",
        "parameters": [
          {
            "in": "query",
            "name": "prescription_id",
            "required": true,
            "schema": {
              "title": "Prescription Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "nudge_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Nudge Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Initiate Doctor Call",
        "tags": [
          "nudges"
        ]
      }
    },
    "/api/v1/nudges/caretaker-tick": {
      "post": {
        "operationId": "caretaker_tick_reminder_api_v1_nudges_caretaker_tick_post",
        "parameters": [
          {
            "in": "query",
            "name": "nudge_id",
            "required": true,
            "schema": {
              "title": "Nudge Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "action",
            "required": false,
            "schema": {
              "default": "checked",
              "enum": [
                "checked",
                "remind_user",
                "call_user"
              ],
              "title": "Action",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Caretaker Tick Reminder",
        "tags": [
          "nudges"
        ]
      }
    },
    "/api/v1/nudges/check-expiry-alerts": {
      "post": {
        "operationId": "trigger_expiry_check_api_v1_nudges_check_expiry_alerts_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Trigger Expiry Check",
        "tags": [
          "nudges"
        ]
      }
    },
    "/api/v1/nudges/cleanup-old": {
      "delete": {
        "operationId": "cleanup_old_nudges_api_v1_nudges_cleanup_old_delete",
        "parameters": [
          {
            "in": "query",
            "name": "before_date",
            "required": false,
            "schema": {
              "title": "Before Date",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 500,
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "dry_run",
            "required": false,
            "schema": {
              "default": true,
              "title": "Dry Run",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "confirm",
            "required": false,
            "schema": {
              "default": "",
              "title": "Confirm",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Cleanup Old Nudges",
        "tags": [
          "nudges"
        ]
      }
    },
    "/api/v1/nudges/doctor-info/{prescription_id}": {
      "get": {
        "operationId": "get_doctor_info_api_v1_nudges_doctor_info__prescription_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "prescription_id",
            "required": true,
            "schema": {
              "title": "Prescription Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Doctor Info",
        "tags": [
          "nudges"
        ]
      }
    },
    "/api/v1/nudges/expiry-alerts": {
      "get": {
        "operationId": "get_expiry_alerts_api_v1_nudges_expiry_alerts_get",
        "parameters": [
          {
            "in": "query",
            "name": "caretaker_id",
            "required": true,
            "schema": {
              "title": "Caretaker Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Expiry Alerts",
        "tags": [
          "nudges"
        ]
      }
    },
...
    "/api/v1/nudges/inbox/caretaker": {
      "get": {
        "operationId": "caretaker_inbox_api_v1_nudges_inbox_caretaker_get",
        "parameters": [
          {
            "in": "query",
            "name": "caretaker_id",
            "required": true,
            "schema": {
              "title": "Caretaker Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "date_filter",
            "required": false,
            "schema": {
              "title": "Date Filter",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Caretaker Inbox",
        "tags": [
          "nudges"
        ]
      }
    },
    "/api/v1/nudges/inbox/user": {
      "get": {
        "operationId": "user_inbox_api_v1_nudges_inbox_user_get",
        "parameters": [
          {
            "in": "query",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "date_filter",
            "required": false,
            "schema": {
              "title": "Date Filter",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "User Inbox",
        "tags": [
          "nudges"
        ]
      }
    },
    "/api/v1/nudges/respond": {
      "post": {
        "operationId": "respond_to_nudge_api_v1_nudges_respond_post",
        "parameters": [
          {
            "in": "query",
            "name": "nudge_id",
            "required": true,
            "schema": {
              "title": "Nudge Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "response",
            "required": false,
            "schema": {
              "default": "taken",
              "enum": [
                "taken",
                "skipped",
                "call_user",
                "checked",
                "remind_user",
                "call_caretaker"
              ],
              "title": "Response",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Respond To Nudge",
        "tags": [
          "nudges"
        ]
      }
    },
    "/api/v1/nudges/send": {
      "post": {
        "operationId": "send_nudge_api_v1_nudges_send_post",
        "parameters": [
          {
            "in": "query",
            "name": "user_id",
            "required": true,
            "schema": {
              "title": "User Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "caretaker_id",
            "required": true,
            "schema": {
              "title": "Caretaker Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "time_bucket",
            "required": true,
            "schema": {
              "enum": [
                "morning",
                "afternoon",
                "night"
              ],
              "title": "Time Bucket",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "message",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Message"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Send Nudge",
        "tags": [
          "nudges"
        ]
      }
    },
    "/api/v1/nudges/sync-daily": {
      "post": {
        "operationId": "sync_daily_nudges_api_v1_nudges_sync_daily_post",
        "parameters": [
          {
            "in": "query",
            "name": "caretaker_id",
            "required": true,
            "schema": {
              "title": "Caretaker Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "user_id",
            "required": false,
            "schema": {
              "default": "",
              "title": "User Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "for_date",
            "required": false,
            "schema": {
              "title": "For Date",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Sync Daily Nudges",
        "tags": [
          "nudges"
        ]
      }
    },
    "/api/v1/nudges/today-status": {
      "get": {
        "operationId": "get_today_status_api_v1_nudges_today_status_get",
        "parameters": [
          {
            "description": "Caretaker user id",
            "in": "query",
            "name": "caretaker_id",
            "required": true,
            "schema": {
              "description": "Caretaker user id",
              "title": "Caretaker Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Today Status",
        "tags": [
          "nudges"
        ]
      }
    },
...
  }
}